% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{catetmle}
\alias{catetmle}
\title{Deprecated, please update code.}
\usage{
catetmle(...)
}
\arguments{
\item{...}{All arguments passed through to new function.}
}
\description{
Deprecated, please update code.
}
