% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{.transport_ate_control}
\alias{.transport_ate_control}
\title{Specify Control Parameters for Estimating the Transport ATE}
\usage{
.transport_ate_control()
}
\description{
Specify Control Parameters for Estimating the Transport ATE
}
