% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_psi_prob.R
\name{est_psi_prob}
\alias{est_psi_prob}
\title{Tests if A is used in the regression function, i.e. if
E[Y|A,W] = E[Y|W] almost surely, or equivalently (under positivity) if
E[Y|A=1,W] = E[Y|A=0,W] almost surely
Returns an estimate of Psi, and a p-value
If est.g is false, then just uses predefined function g0
NOTE: Theory all uses bounded Y, so if Y is not bounded then at least try to make sure most of its
mass falls in the interval [-1,1] --- can do this by scaling by a constant}
\usage{
est_psi_prob(W, A, Y, W.train = NULL, A.train = NULL, Y.train = NULL,
  sig.meth = "var", est.g = TRUE, g0 = NULL, SL.library = c("SL.rpart",
  "SL.glm", "SL.earth", "SL.nnet", "SL.gam"))
}
\description{
Tests if A is used in the regression function, i.e. if
E[Y|A,W] = E[Y|W] almost surely, or equivalently (under positivity) if
E[Y|A=1,W] = E[Y|A=0,W] almost surely
Returns an estimate of Psi, and a p-value
If est.g is false, then just uses predefined function g0
NOTE: Theory all uses bounded Y, so if Y is not bounded then at least try to make sure most of its
mass falls in the interval [-1,1] --- can do this by scaling by a constant
}
\references{
Luedtke et al. 2017
}
\author{
Alex Luedtke
}
