% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_psi_prob_binom.R
\name{est_psi_prob_binom}
\alias{est_psi_prob_binom}
\title{Omnibus test}
\usage{
est_psi_prob_binom(W, A, Y, W.train = NULL, A.train = NULL,
  Y.train = NULL, sig.meth = "var", est.g = TRUE, g0 = NULL,
  SL.library = c("SL.glm", "SL.step", "SL.earth", "SL.gam", "SL.nnet",
  "SL.mean", "SL.glm.interaction"))
}
\description{
Omnibus test
}
\references{
Luedtke et al. 2017
}
\author{
Alex Luedtke
}
