% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{ittatetmle}
\alias{ittatetmle}
\title{Deprecated, please update code.}
\usage{
ittatetmle(...)
}
\arguments{
\item{...}{All arguments passed through to new function.}
}
\description{
Deprecated, please update code.
}
