% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{ittatetmle_multi_z}
\alias{ittatetmle_multi_z}
\title{fit ittate_TMLE with multi-dimensional Z}
\usage{
ittatetmle_multi_z(a, z, y, site, w, aamodel, asitemodel, s_awz_model,
  s_aw_model, aoutmodel, aq2model)
}
\arguments{
\item{a}{needs to be named a and have values 0/1}

\item{z}{matrix, each column be an individual z covariate}

\item{y}{{ parameter_description }}

\item{site}{The site}

\item{w}{{ parameter_description }}

\item{aamodel}{The aamodel}

\item{asitemodel}{The asitemodel}

\item{s_awz_model}{formula for fitting S ~ A + W + Z}

\item{s_aw_model}{formula for fitting S ~ A + W}

\item{aoutmodel}{The aoutmodel}

\item{aq2model}{The aq 2 model}

\item{weights}{vector with length == nrow(data). p(Delta_2 | A,W,S)}
}
\value{
{ description_of_the_return_value }
}
\description{
calculates the ratio dga1s0/dga1s1 with a trick in Zheng(2012)
}
