% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_SL.R
\name{ittatetmle_multi_z_sl}
\alias{ittatetmle_multi_z_sl}
\title{fit ittate_TMLE with SuperLearner}
\usage{
ittatetmle_multi_z_sl(a, z, y, site, w, aamodel, asitemodel, s_awz_model,
  s_aw_model, aoutmodel, aq2model)
}
\description{
fit ittate_TMLE with SuperLearner
}
