% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd_test.R
\name{mmd_test}
\alias{mmd_test}
\title{Test of maximum mean discrepancy}
\usage{
mmd_test(R, S, D.R, D.S, sig.meth = "eig", num.reps = 10000,
  return.cutoff = FALSE)
}
\description{
Test of maximum mean discrepancy
}
\examples{

# TBD.

}
\references{
Luedtke et al. 2017
}
\author{
Alex Luedtke
}
