% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notransport.R
\name{notransport}
\alias{notransport}
\title{No-transport estimator}
\usage{
notransport(a, z, w, site, ntamodel, ntzmodel)
}
\description{
This estimator is also used in \code{\link{transport_cace}}.
}
\details{
<Parameter info>

<Return info>
}
\examples{

# TBD.

}
\references{
TBD.
}
