\name{random}
\docType{data}
\alias{random32}
\alias{random64}
\alias{random128}
\alias{random32a}
\alias{random32b}
\alias{random64a}
\alias{random64b}
\alias{random128a}
\alias{random128b}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Images to Illustrate the Use of transport.pgrid
}
\description{
32 x 32, 64 x 64 and 128 x 128 images to illustrate the use of \code{\link{transport.pgrid}}. These are objects of 
class \code{"pgrid"}.
}
\usage{
random32a

random32b

random64a

random64b

random128a

random128b
}
%- maybe also 'usage' for other objects documented here.
\format{
  Objects of class \sQuote{pgrid}.
}
\source{
  Randomly generated using the package \code{RandomFields}.
}
\keyword{datasets}