% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semidiscrete1.R
\name{ret_message}
\alias{ret_message}
\title{Return Text Strings for lbfgs Return Codes}
\usage{
ret_message(n = NULL)
}
\arguments{
\item{n}{The vector of return codes or \code{NULL} meaning the whole list
shall be returned.}
}
\value{
A named character vector of the corresponding return strings.
}
\description{
Given a vector of return codes, give back the corresponding vector of 
return strings from the lbfgs library. Nonexistant codes are ignored.
}
\note{
Code 0 is ignored, since for technical reasons it is never returned by
      the function \code{\link{semidiscrete1}}.
}
\examples{
ret_message()
ret_message(c(2,-1023,-1019))

}
\seealso{
\code{\link{semidiscrete1}}.
}
\author{
Dominic Schuhmacher \email{schuhmacher@math.uni-goettingen.de}
}
\keyword{code}
\keyword{return}
