% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_ate.R
\name{transport_ate}
\alias{transport_ate}
\title{Transport Estimator for the Average Treatment Effect}
\usage{
transport_ate(
  data,
  trt,
  outcome,
  source,
  covar,
  cens = NULL,
  outcome_type = c("binomial", "continuous"),
  id = NULL,
  weights = NULL,
  control = .transport_ate_control()
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A \code{data.frame} in wide format containing all necessary variables
for the estimation problem.}

\item{trt}{[\code{character(1)}]\cr
The column name of the treatment variable. This variable should be coded as 0 or 1.}

\item{outcome}{[\code{character(1)}]\cr
The column name of the outcome variable.}

\item{source}{[\code{character(1)}]\cr
The column name of the population indicator. This variable should be coded as 0 or 1.}

\item{covar}{[\code{character}]\cr
An vector containing the column names of baseline covariates to be
included for adjustment.}

\item{cens}{[\code{character}]\cr
An optional column name of a censoring indicator. If missingness in the outcome is
present, must be provided. CURRENTLY IGNORED.}

\item{outcome_type}{[\code{character(1)}]\cr
Outcome variable type (i.e., continuous, binomial).}

\item{id}{[\code{character(1)}]\cr
An optional column name containing cluster level identifiers. CURRENTLY IGNORED.}

\item{weights}{[\code{numeric(nrow(data))}]\cr
An optional vector containing sampling weights. CURRENTLY IGNORED.}

\item{control}{[\code{list}]\cr See \code{.transport_ate_control()}.}
}
\description{
One-step transport estimator for the average treatment effect (ATE)
under the general model that allows for all confounders to modify
the ATE.
}
\seealso{
[.transport_ate_control()].
}
