% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_cace.R
\name{transport_cace}
\alias{transport_cace}
\title{Transported complier average causal effect (CACE) (aka treatment effect on treated)}
\usage{
transport_cace(ca, cz, cy, csite, cw, czmodel, csitemodel, coutmodel, cq2model)
}
\description{
Details to be added.
}
\details{
<Add parameter info>

<Add return info>
}
\examples{

# TBD.

}
\references{
Imbens and Rubin, 1997
}
