% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_eace.R
\name{transport_eace}
\alias{transport_eace}
\title{Average causal effect of the exposure on the outcome, ignoring the instrument}
\usage{
transport_eace(a, z, y, site, w, nsitemodel, nzmodel, noutmodel)
}
\description{
Average causal effect of the exposure on the outcome, ignoring the instrument
}
\examples{

# TBD.

}
\references{
TBD.
}
