% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{TransportTask}
\alias{TransportTask}
\title{Transport Task Class}
\description{
Transport Task Class

Transport Task Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransportTask-new}{\code{TransportTask$new()}}
\item \href{#method-TransportTask-print}{\code{TransportTask$print()}}
\item \href{#method-TransportTask-training}{\code{TransportTask$training()}}
\item \href{#method-TransportTask-validation}{\code{TransportTask$validation()}}
\item \href{#method-TransportTask-select}{\code{TransportTask$select()}}
\item \href{#method-TransportTask-observed}{\code{TransportTask$observed()}}
\item \href{#method-TransportTask-clone}{\code{TransportTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$new(
  data,
  trt,
  outcome,
  covar,
  pop,
  modifiers = NULL,
  obs = NULL,
  group = NULL,
  weights = NULL,
  folds = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-print"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-training"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-training}{}}}
\subsection{Method \code{training()}}{
Get training and validation sets for a fold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$training(fold)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fold}}{Fold number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-validation"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-validation}{}}}
\subsection{Method \code{validation()}}{
Get training and validation sets for a fold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$validation(fold)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fold}}{Fold number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-select"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-select}{}}}
\subsection{Method \code{select()}}{
Subset and return the task data, keeping only features from `cols`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$select(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{A character vector of column names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-observed"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-observed}{}}}
\subsection{Method \code{observed()}}{
Subset and return the task data, keeping only features from `cols` that aren't censored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$observed(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{A character vector of column names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransportTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransportTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransportTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
