\name{methods}
\alias{print.pgrid}
\alias{print.pp}
\alias{print.wpp}
\alias{summary.pgrid}
\alias{summary.pp}
\alias{summary.wpp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Print and Summary Methods for Objects of Class pgrid, pp and wpp
}
\description{
  Prints a brief description of a pixel grid or a point pattern.
}
\usage{
\method{print}{pgrid}(x, ...)
\method{print}{pp}(x, ...)
\method{print}{wpp}(x, ...)
\method{summary}{pgrid}(object, ...)
\method{summary}{pp}(object, ...)
\method{summary}{wpp}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x, object}{
    an object of class pgrid or pp or wpp.
  }
  \item{...}{
  	additional arguments. Currently without effect.
  }
}
\details{
  Currently there is no difference between print and summary.
}
\author{
  Dominic Schuhmacher \email{dschuhm1@uni-goettingen.de} \cr
  Timo Wilm \email{timo.wilm@stud.uni-goettingen.de}
}

