% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-package.R
\docType{package}
\name{transport-package}
\alias{transport}
\alias{transport-package}
\title{transport: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
