% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_ate.R
\name{transport_ate}
\alias{transport_ate}
\title{Title}
\usage{
transport_ate(
  data,
  trt,
  outcome,
  covar,
  pop,
  obs = NULL,
  id = NULL,
  weights = NULL,
  estimator = c("standard", "collaborative"),
  learners_trt = "glm",
  learners_pop = "glm",
  learners_outcome = "glm",
  learners_heterogeneity = "glm",
  folds = 1,
  control = transport_control()
)
}
\arguments{
\item{control}{}
}
\description{
Title
}
