% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_control.R
\name{transport_control}
\alias{transport_control}
\title{Title}
\usage{
transport_control(
  .learners_folds = NULL,
  .bound = 1e+05,
  .return_full_fits = FALSE,
  .discrete = FALSE,
  .info = FALSE
)
}
\arguments{
\item{.info}{}
}
\description{
Title
}
